ready

+++ statistics-service/statistics-server/src/main/java/ru/practicum/model/StatisticView.java
@@ -6,6 +6,6 @@ public interface StatisticView {
 
     String getUri();
 
-    Long getHits();
+    Integer getHits();
 
 }
